requirement.txt
cv2
pyfiglet
numpy
Pillow
halo